smart Ai cheating detection/requirements.txt
Flask
mysql-connector-python
opencv-python
opencv-contrib-python
mediapipe
numpy
scikit-learn
SpeechRecognition
Pillow
Werkzeug